Clean and restructure files
