+++ src/components/AssetManagement/Checkout/CheckoutSubmitSuccess.js
