+++ src/components/Resources/Landing/ResourcesLandingTabsGroups.js
@@ -1,19 +1,19 @@
 import { Link } from 'react-router-dom';
 import React from 'react';
-import tools from "../../icons/resources/group-category-icons/tools.svg";
-import apple from "../../icons/resources/group-category-icons/apple.svg";
-import clock from "../../icons/resources/group-category-icons/clock.svg";
-import document from "../../icons/resources/group-category-icons/document.svg";
-import fishbowl from "../../icons/resources/group-category-icons/fishbowl.svg";
-import globe from "../../icons/resources/group-category-icons/globe.svg";
-import group from "../../icons/resources/group-category-icons/group.svg";
-import infobook from "../../icons/resources/group-category-icons/infobook.svg";
-import tdx from "../../icons/resources/group-category-icons/tdx.svg";
-import shield from "../../icons/resources/group-category-icons/shield.svg";
-import slasheye from '../../icons/resources/group-category-icons/slasheye.svg'
-import umich from '../../icons/resources/group-category-icons/umich.svg'
-import printer from "../../icons/resources/group-category-icons/printer.svg";
-import wrench from "../../icons/resources/group-category-icons/wrench.svg";
+import tools from "../../../icons/resources/group-category-icons/tools.svg";
+import apple from "../../../icons/resources/group-category-icons/apple.svg";
+import clock from "../../../icons/resources/group-category-icons/clock.svg";
+import document from "../../../icons/resources/group-category-icons/document.svg";
+import fishbowl from "../../../icons/resources/group-category-icons/fishbowl.svg";
+import globe from "../../../icons/resources/group-category-icons/globe.svg";
+import group from "../../../icons/resources/group-category-icons/group.svg";
+import infobook from "../../../icons/resources/group-category-icons/infobook.svg";
+import tdx from "../../../icons/resources/group-category-icons/tdx.svg";
+import shield from "../../../icons/resources/group-category-icons/shield.svg";
+import slasheye from '../../../icons/resources/group-category-icons/slasheye.svg'
+import umich from '../../../icons/resources/group-category-icons/umich.svg'
+import printer from "../../../icons/resources/group-category-icons/printer.svg";
+import wrench from "../../../icons/resources/group-category-icons/wrench.svg";
 
 const iconMapping = {
     apple: apple,
